Replace Authorizer with Token concept
