+++ README.md
@@ -2,7 +2,7 @@
 
 # JWT::Authorizer
 
-`JWT::Authorizer` makes authorization with [JWT tokens](https://jwt.io/) simple. It allows creating and verifying JWT tokens according to rules (validations) set on specific `Authorizer` class.
+`JWT::Authorizer` makes authorization with [JWT tokens](https://jwt.io/) simple. It allows creating and verifying JWT tokens according to claims set on specific `JWT::Token` class.
 
 ## Installation
 
@@ -24,19 +24,19 @@ Or install it yourself as:
 
 ### Configuration
 
-You can configure your `JWT::Authorizer` classes with `.configuration` and `.configure` options:
+You can configure your `JWT::Token` classes with `.configuration` and `.configure` options:
 
 ```ruby
-JWT::Authorizer.configuration
+JWT::Token.configuration
 
-JWT::Authorizer.configure do |config|
+JWT::Token.configure do |config|
   config.expiry = 12 * 60 * 60
   config.algorithm = "RS256"
   config.secret = { private_key: nil, public_key: ENV["SECRET_KEY"] }
 end
 ```
 
-`JWT::Authorizer` have following options available:
+`JWT::Token` have following options available:
 
 * `algorithm` - determines algorithm used on signing and verifying JWT tokens. Defaults to `"HS256"`.
 * `secret` - for [`HMAC`](https://en.wikipedia.org/wiki/HMAC) algorithms it accepts simple `String` with symmetric key, for [`RSA`](https://en.wikipedia.org/wiki/RSA_(cryptosystem)) and [`ECDSA`](https://en.wikipedia.org/wiki/Elliptic_Curve_Digital_Signature_Algorithm) it requires hash with `:private_key` and `:public_key` keys.
@@ -44,35 +44,35 @@ end
 * `issuer` - sets `iss` claim in the token. Defaults to `nil`.
 * `allowed_issuers` - array of issuers that will be allowed on token verification. Defaults to empty array, tokens with any value in `iss` claim (and without this claim) will be valid. If array contains any elements, *only* listed issuers will be valid.
 
-Default options can be overriden during instantiation of `JWT::Authorizer` classes:
+Default options can be overriden during instantiation of `JWT::Token` classes:
 
 ```ruby
-JWT::Authorizer.configuration.expiry #=> 3600
-JWT::Authorizer.new(expiry: 60).expiry #=> 60
+JWT::Token.configuration.expiry #=> 3600
+JWT::Token.new(expiry: 60).expiry #=> 60
 ```
 
 ### Generating tokens
 
-To generate JWT token, create instance of `JWT::Authorizer` and call `#build` method. It accepts hash of additional claims you want in your token.
+To generate JWT token, create instance of `JWT::Token` and call `#build` method. It accepts hash of additional claims you want in your token.
 
 ```ruby
-JWT::Authorizer.configuration.secret = "hmac"
-JWT::Authorizer.new.build(level: :admin)
+JWT::Token.configuration.secret = "hmac"
+JWT::Token.new.build(level: :admin)
 #=> "eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjAyODQ3MTcsImxldmVsIjoiYWRtaW4ifQ.nHRIBBjzteHuzygij-BlfXx3YIvfeO39Qh84hq729KQ"
 ```
 
 ### Verifying tokens
 
-To verify token, use `JWT::Authorizer#verify` method.
+To verify token, use `JWT::Token#verify` method.
 
 ```ruby
-JWT::Authorizer.configuration.secret = "hmac"
+JWT::Token.configuration.secret = "hmac"
 token = "eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjAyODUwMzd9.CO8K_mqXCZfu8W12tpYcBo1WyrLZAmEMmr8R-HM3a5E"
-JWT::Authorizer.new.verify(token)
+JWT::Token.new.verify(token)
 #=> [{"exp"=>1520285037}, {"alg"=>"HS256"}]
-JWT::Authorizer.new.verify(nil)
+JWT::Token.new.verify(nil)
 # JWT::DecodeError: Nil JSON web token
-JWT::Authorizer.new.verify("eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjB9.nooope")
+JWT::Token.new.verify("eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjB9.nooope")
 # JWT::VerificationError: Signature verification raised
 ```
 
@@ -81,7 +81,7 @@ JWT::Authorizer.new.verify("eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjB9.nooope")
 You can use validators to verify non-standard claims.
 
 ```ruby
-class AdminAuthorizer < JWT::Authorizer
+class AdminAuthorizer < JWT::Token
   validate :level, required: true do |value, _context|
     raise JWT::DecodeError, "Level must be admin" unless value == "admin"
   end
@@ -92,7 +92,7 @@ AdminAuthorizer.new.verify(valid_token)
 # [{"exp"=>1520285339, "level"=>"admin"}, {"alg"=>"HS256"}]
 missing_claim = "eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjAyODUzODd9.ncXmy81O64OjLNP4eCdAyVklAfGqdYiWp0K6FoI1pec"
 AdminAuthorizer.new.verify(missing_claim)
-# JWT::Authorizer::MissingClaim: Token is missing required claim: level
+# JWT::Token::MissingClaim: Token is missing required claim: level
 invalid_value = "eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjAyODU0MzQsImxldmVsIjoicmVndWxhciJ9.z16nhJcOpRJmDZdkrDrdo1TetQ9YZpYiQmBdc53lnV0"
 AdminAuthorizer.new.verify(invalid_value)
 # JWT::DecodeError: Level must be admin
@@ -103,7 +103,7 @@ AdminAuthorizer.new.verify(invalid_value)
 You can pass additional context to validators:
 
 ```ruby
-class AdminAuthorizer < JWT::Authorizer
+class AdminAuthorizer < JWT::Token
   validate :path do |value, rack_request|
     raise JWT::DecodeError, "invalid path" unless value == rack_request.path
   end
@@ -112,7 +112,7 @@ end
 AdminAuthorizer.new.verify(token, rack_request)
 ```
 
-See [`JWT::RequestAuthorizer`](lib/jwt/request_authorizer.rb) and it's [spec](spec/jwt/request_authorizer_spec.rb) for examples.
+See [`JWT::EndpointToken`](lib/jwt/endpoint_token.rb) and it's [spec](spec/jwt/endpoint_token_spec.rb) for examples.
 
 ## Development
 
@@ -130,4 +130,4 @@ The gem is available as open source under the terms of the [MIT License](https:/
 
 ## Code of Conduct
 
-Everyone interacting in the JWT::Authorizer project’s codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/[USERNAME]/jwt-authorizer/blob/master/CODE_OF_CONDUCT.md).
+Everyone interacting in the JWT::Token project’s codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/[USERNAME]/jwt-authorizer/blob/master/CODE_OF_CONDUCT.md).
