+++ lib/jwt/authorizer/version.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module JWT
-  class Authorizer
+  module Authorizer
     VERSION = "1.0.0.beta1"
   end
 end
